requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.0+cpu
torchvision==0.17.0+cpu
numpy<2.0.0

streamlit
langchain-huggingface
langchain-community
chromadb
sentence-transformers
pypdf
openai
python-dotenv
requests
pillow
markdown
pdfkit
transformers==4.48.2
